made the template delete feature work with a simple javascript confirm

+++ www/sim-template-edit.adp
@@ -5,11 +5,12 @@
 
 <formtemplate id="sim_template"></formtemplate>
 
-<p>
-  <b>&raquo;</b> <a href="@delete_url@">Delete this template</a>
-</p>
-
 <if @workflow_id@ not nil>
+
+  <p>
+    <b>&raquo;</b> <a href="@delete_url@" onclick="return confirm('Are you sure you want to delete the template?');">Delete this template</a>
+  </p>
+
   <h4>Roles</h4>
 
   <include src="/packages/simulation/lib/sim-template-roles" workflow_id="@workflow_id@" display_mode="edit">
